requirements.txt
django
pillow
Faker